Refactored main to use more final architecture.
main() is much simpler looking now, with the while(running) loop replaced by just two clean lines.
First, an instruction is fetched from the program at the specified position of the program counter.
Then, it is evaluated--which for now means 'check to see if it's a halt instruction.'
Then we step to the next instruction. The loop ends if we hit a halt.